[OkHttp] Divide base class into bodyFrom and bodyTo okhttp
